% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_build_goal_events.R
\name{classify_goal_type}
\alias{classify_goal_type}
\title{Map a free-text "Type of goal" value to the vocabulary in
config.yaml's goal_type_taxonomy. Anything not listed there is an error,
not silently dropped -- a new Transfermarkt label must be classified, not
guessed at.}
\usage{
classify_goal_type(goal_type, taxonomy)
}
\arguments{
\item{goal_type}{Free-text "Type of goal" value from the Transfermarkt log.}

\item{taxonomy}{The \code{goal_type_taxonomy} list from config.yaml.}
}
\value{
See description above.
}
\description{
Map a free-text "Type of goal" value to the vocabulary in
config.yaml's goal_type_taxonomy. Anything not listed there is an error,
not silently dropped -- a new Transfermarkt label must be classified, not
guessed at.
}
