% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ingest_wikipedia.R
\name{clean_wiki_markup}
\alias{clean_wiki_markup}
\title{Strip the wiki markup this table uses around player names -- {{flagicon}}
and {{efn}} templates, italics for still-active players -- then resolve
the remaining wikilink to display text.}
\usage{
clean_wiki_markup(x)
}
\arguments{
\item{x}{Text containing wiki markup ({{templates}}, [[wikilinks]]).}
}
\value{
See description above.
}
\description{
Strip the wiki markup this table uses around player names -- {{flagicon}}
and {{efn}} templates, italics for still-active players -- then resolve
the remaining wikilink to display text.
}
