% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_insights.R
\name{contemporary_insight}
\alias{contemporary_insight}
\title{Seasons Shearer topped or ran up the league's scoring charts.}
\usage{
contemporary_insight(season_summary, contemporary_summary)
}
\arguments{
\item{season_summary}{Output of \code{season_summary()}.}

\item{contemporary_summary}{Output of \code{contemporary_summary()}.}
}
\value{
A glue string.
}
\description{
Seasons Shearer topped or ran up the league's scoring charts.
}
