% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_figures.R
\name{fig_context}
\alias{fig_context}
\title{Bars of goal context (equaliser / go_ahead / reduce_deficit / extend_lead).}
\usage{
fig_context(context_breakdown)
}
\arguments{
\item{context_breakdown}{Output of \code{context_breakdown()}.}
}
\value{
A ggplot (or gt table) object.
}
\description{
Bars of goal context (equaliser / go_ahead / reduce_deficit / extend_lead).
}
