% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_figures.R
\name{fig_goal_types}
\alias{fig_goal_types}
\title{Stacked bar of goal-type shares, "unknown" shown alongside classified types.}
\usage{
fig_goal_types(goal_type_breakdown)
}
\arguments{
\item{goal_type_breakdown}{Output of \code{goal_type_breakdown()}.}
}
\value{
A ggplot (or gt table) object.
}
\description{
Stacked bar of goal-type shares, "unknown" shown alongside classified types.
}
