% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_build_matches.R
\name{football_data_season_label}
\alias{football_data_season_label}
\title{football-data.co.uk season codes ("9394") to season labels ("1993-94").}
\usage{
football_data_season_label(code)
}
\arguments{
\item{code}{A football-data.co.uk season code, e.g. "9394".}
}
\value{
See description above.
}
\description{
football-data.co.uk season codes ("9394") to season labels ("1993-94").
}
