% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_sources.R
\name{load_config}
\alias{load_config}
\title{Load config/config.yaml as a nested list.}
\usage{
load_config(path = "config/config.yaml")
}
\arguments{
\item{path}{Output file path.}
}
\value{
See description above.
}
\description{
Load config/config.yaml as a nested list.
}
