% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_analyse.R
\name{penalty_summary}
\alias{penalty_summary}
\title{Penalty goals as a share of the career total.}
\usage{
penalty_summary(goal_events)
}
\arguments{
\item{goal_events}{The \code{shearer_goal_events} tibble (goal_events_checked in \verb{_targets.R}).}
}
\value{
See description above.
}
\description{
Penalty goals as a share of the career total.
}
