% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_shared.R
\docType{data}
\name{SHEARER_COLOUR}
\alias{SHEARER_COLOUR}
\title{Newcastle/Blackburn-adjacent red, used only for Shearer.}
\format{
An object of class \code{character} of length 1.
}
\usage{
SHEARER_COLOUR
}
\description{
Newcastle/Blackburn-adjacent red, used only for Shearer.
}
\keyword{datasets}
